Add files via upload
in this .sql file, I implemented normalization. this is a restaurant, that is capable of delivering food to customers. several tables have been created to ensure there is no redundancy and every table is used in an efficient manner, updating or deleting user data will not affect the entire database.